whisper-server/requirements.txt
openai-whisper
flask
flask-cors
